Add files via upload
This is Data set to Classify the Benign and Malignant cells in the given data set using the description about the cells in the form of columnar attributes. The breast cancer data includes 569 examples of cancer biopsies, each with 32 features. One feature is an identification number, another is the cancer diagnosis and 30 are numeric-valued laboratory measurements. The diagnosis is coded as "M" to indicate malignant or "B" to indicate benign.